+++ app/Http/Controllers/MentorController.php
@@ -14,6 +14,9 @@ class MentorController extends Controller
      */
     public function showAll()
     {
+        if (!auth()->user()->hasPermissionTo('see_mentors')) {
+            return response()->json(['Message' => "Action non-autorisée"], 401);
+        }
         $mentors = Mentor::all();
         foreach ($mentors as $mentor) {
             $mentor = $mentor->applicant->user;
@@ -29,6 +32,9 @@ class MentorController extends Controller
      */
     public function showOne($id)
     {
+        if (!auth()->user()->hasPermissionTo('see_mentor')) {
+            return response()->json(['Message' => "Action non-autorisée"], 401);
+        }
         $mentor = Mentor::findOrFail($id);
         $mentor = $mentor->applicant->user;
         return response()->json($mentor);
@@ -36,6 +42,9 @@ class MentorController extends Controller
 
     public function update($id, Request $request)
     {
+        if (!auth()->user()->hasPermissionTo('update_mentor')) {
+            return response()->json(['Message' => "Action non-autorisée"], 401);
+        }
         $mentor = Mentor::findOrFail($id);
         $mentor->update($request->all());
         $mentor->applicant->update($request->all());
@@ -45,18 +54,27 @@ class MentorController extends Controller
 
     public function delete($id)
     {
+        if (!auth()->user()->hasPermissionTo('delete_mentor')) {
+            return response()->json(['Message' => "Action non-autorisée"], 401);
+        }
         Mentor::findOrFail($id)->delete();
         return response('Supprimé avec succès!', 200);
     }
 
     public function getSkill($id)
     {
+        if (!auth()->user()->hasPermissionTo('see_mentor_skills')) {
+            return response()->json(['Message' => "Action non-autorisée"], 401);
+        }
         $mentor = Mentor::findOrFail($id);
         return response()->json($mentor->applicant->skills, 200);
     }
 
     public function setSkill($id, Request $request)
     {
+        if (!auth()->user()->hasPermissionTo('add_mentor_skill')) {
+            return response()->json(['Message' => "Action non-autorisée"], 401);
+        }
         $mentor = Mentor::findOrFail($id);
         $mentor->applicant->skills()->attach($request->skill);
         return response()->json($mentor->applicant->skills, 200);
@@ -64,6 +82,9 @@ class MentorController extends Controller
 
     public function deleteSkill($id, Request $request)
     {
+        if (!auth()->user()->hasPermissionTo('delete_mentor_skill')) {
+            return response()->json(['Message' => "Action non-autorisée"], 401);
+        }
         $mentor = Mentor::findOrFail($id);
         $mentor->applicant->skills()->detach($request->skill);
         return response()->json($mentor->applicant->skills, 200);
@@ -71,12 +92,18 @@ class MentorController extends Controller
 
     public function getArea($id)
     {
+        if (!auth()->user()->hasPermissionTo('see_mentor_area')) {
+            return response()->json(['Message' => "Action non-autorisée"], 401);
+        }
         $mentor = Mentor::findOrFail($id);
         return response()->json($mentor->applicant->areas, 200);
     }
 
     public function setArea($id, Request $request)
     {
+        if (!auth()->user()->hasPermissionTo('add_mentor_area')) {
+            return response()->json(['Message' => "Action non-autorisée"], 401);
+        }
         $mentor = Mentor::findOrFail($id);
         $mentor->applicant->areas()->attach($request->area);
         return response()->json($mentor->applicant->areas, 200);
@@ -84,6 +111,9 @@ class MentorController extends Controller
 
     public function deleteArea($id, Request $request)
     {
+        if (!auth()->user()->hasPermissionTo('delete_mentor_area')) {
+            return response()->json(['Message' => "Action non-autorisée"], 401);
+        }
         $mentor = Mentor::findOrFail($id);
         $mentor->applicant->areas()->detach($request->area);
         return response()->json($mentor->applicant->areas, 200);
